# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ShopEasy Gateway API
  version: 1.0.0
  description: |
    OpenAPI description for the gateway-facing ShopEasy API.

    Notes:
    - Browser clients must fetch `GET /api/csrf-token` and send `X-CSRF-Token` on mutating requests.
    - Authenticated routes use `Authorization: Bearer <access_token>`.
    - Guest cart routes use `X-Cart-ID` when a guest cart has already been created.
servers:
  - url: http://localhost
    description: Local storefront + gateway via Nginx
  - url: http://localhost:5000
    description: Direct gateway access
...
paths:
...
  /api/cart:
    get:
      tags: [Cart]
      summary: Get the current guest or authenticated cart
      parameters:
        - $ref: '#/components/parameters/XCartId'
      responses:
        '200':
          description: Cart payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
  /api/cart/add/{product_id}:
    post:
      tags: [Cart]
      summary: Add a product to the cart
      security:
        - csrfToken: []
      parameters:
        - in: path
          name: product_id
          required: true
          schema:
            type: integer
        - $ref: '#/components/parameters/XCartId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                quantity:
                  type: integer
                  minimum: 1
      responses:
        '200':
          description: Item added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartMutationResponse'
        '403':
          description: Missing or invalid CSRF token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/cart/remove/{product_id}:
    post:
      tags: [Cart]
      summary: Remove quantity or remove an item from the cart
      security:
        - csrfToken: []
      parameters:
        - in: path
          name: product_id
          required: true
          schema:
            type: integer
        - $ref: '#/components/parameters/XCartId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                quantity:
                  type: integer
                  minimum: 1
      responses:
        '200':
          description: Item removed or reduced
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  cart_count:
                    type: integer
                required: [success, cart_count]
  /api/cart/clear:
    post:
      tags: [Cart]
      summary: Clear the current cart
      security:
        - csrfToken: []
      parameters:
        - $ref: '#/components/parameters/XCartId'
      responses:
        '200':
          description: Cart cleared
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                required: [success]
  /api/cart/count:
    get:
      tags: [Cart]
      summary: Get total item count in the cart
      parameters:
        - $ref: '#/components/parameters/XCartId'
      responses:
        '200':
          description: Cart count
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                required: [count]
  /api/cart/merge:
    post:
      tags: [Cart]
      summary: Merge a guest cart into the authenticated user's cart
      security:
        - bearerAuth: []
        - csrfToken: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                guest_cart_id:
                  type: string
                  format: uuid
      responses:
        '200':
          description: Merge result
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  cart_id:
                    type: string
                    format: uuid
                  message:
                    type: string
                required: [success]
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
  parameters:
    XCartId:
      in: header
      name: X-Cart-ID
      required: false
      schema:
        type: string
        format: uuid
      description: Guest cart identifier returned by cart mutation endpoints
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
...
    CartItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        image_url:
          type: string
        category:
          type: string
        stock:
          type: integer
        quantity:
          type: integer
      required: [id, name, price, quantity]
    CartResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CartItem'
        cart_id:
          type: string
          nullable: true
          format: uuid
        total:
          type: number
      required: [items, cart_id, total]
    CartMutationResponse:
      type: object
      properties:
        success:
          type: boolean
        cart_id:
          type: string
          format: uuid
        product_id:
          type: integer
        quantity:
          type: integer
        cart_count:
          type: integer
      required: [success, cart_id, product_id, quantity, cart_count]
...